requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
arrow==1.3.0
async-timeout==4.0.3
attrs==24.2.0
canvasapi==3.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
discord==2.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
propcache==0.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.16.0
